pkp/pkp-lib#4844 Divided the Statistics.vue into two sub-components.
