+++ src/components/Statistics/Statistics.less
@@ -1,0 +1,167 @@
+@import '../../styles/_import';
+
+.pkpStatistics .pkpPageHeader {
+	padding-left: 0;
+	padding-right: 0;
+}
+
+.pkpStatistics .pkpListPanel__filter.-isVisible {
+	float: left;
+	position: relative;
+	left: auto;
+	width: 192px;
+	margin-left: -@base;
+
+	&:before {
+		display: none;
+	}
+
+	+ .pkpStatistics__main  {
+		margin-left: 176px;
+	}
+}
+
+.pkpStatistics__main {
+	margin-left: 0;
+	transition: margin-left 0.2s;
+}
+
+.pkpStatistics__tableHeader {
+	margin-top: 2rem;
+}
+
+.pkpStatistics__tableTitle {
+	display: inline-block;
+	margin: 0;
+	font-size: @font-base;
+	line-height: 1.5em;
+
+	.pkpSpinner {
+		margin-left: 0.5rem;
+	}
+}
+
+.pkpStatistics__tableActions {
+	float: right;
+	margin-left: 1rem;
+}
+
+.pkpStatistics__itemsOfTotal {
+	font-size: @font-tiny;
+}
+
+.pkpStatistics .pkpTable {
+
+	.pkpListPanel__search {
+		display: inline-block;
+		float: none;
+		margin-top: 0;
+		margin-left: 1rem;
+		max-width: 20em;
+	}
+
+	.pkpListPanel__searchInput {
+		font-size: @font-tiny;
+		line-height: 2.5em;
+		padding-left: 2.5rem;
+	}
+}
+
+.pkpStatistics__itemLink {
+	color: @text;
+	text-decoration: none;
+}
+
+.pkpStatistics__itemAuthors {
+	font-weight: @bold;
+}
+
+.pkpStatistics__noRecords {
+	padding: @double @base;
+	border: @grid-border;
+	border-top: none;
+	font-size: @font-sml;
+	text-align: center;
+	color: @text-light;
+}
+
+.pkpStatistics__graph {
+	background: @bg-anchor;
+	color: #fff;
+	border-radius: @radius;
+
+	.chartjs-render-monitor {
+		border-radius: @radius;
+	}
+}
+
+.pkpStatistics__graphHeader {
+	padding: @base;
+}
+
+.pkpStatistics__graphTitle {
+	display: inline-block;
+	margin: 0;
+	font-size: @base;
+	line-height: 1.5em;
+}
+
+.pkpStatistics__graphSegment {
+	float: right;
+
+	.pkpButton {
+		position: relative;
+		z-index: 1;
+		float: right;
+		background: transparent;
+		border: 1px solid #437b96;
+		box-shadow: 0 1px 0 #000;
+		font-size: @font-tiny;
+		line-height: 2em;
+		color: #fff;
+
+		&:before {
+			content: '';
+			position: relative;
+			display: inline-block;
+			width: 0.75em;
+			height: 0.75em;
+			margin-right: 0.25em;
+			border: 1px solid #fff;
+			border-radius: 50%;
+		}
+
+		&:first-child {
+			border-top-left-radius: 0;
+			border-bottom-left-radius: 0;
+		}
+
+		&:last-child {
+			position: relative;
+			left: 1px;
+			border-top-right-radius: 0;
+			border-bottom-right-radius: 0;
+		}
+
+		&:hover,
+		&:focus {
+			color: #fff;
+			border-color: #fff;
+			z-index: 2;
+		}
+
+		&[aria-pressed="true"] {
+			background: @primary;
+
+			&:before {
+				background: #fff;
+				box-shadow: inset 0 0 0 1px @primary;
+			}
+		}
+
+		&[disabled] {
+			background: transparent;
+			opacity: 0.5;
+		}
+	}
+}
